refactor: 에코 보내기 버튼 컴포넌트 분리
